requirements.txt
uvicorn 
fastapi
python-multipart
google-generativeai
python-dotenv
requests
Pillow
